+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [Unreleased]
 
 - Upgrade to `protoc` 3.19.2 support.
-- Add `buf beta registry {plugin,template} {deprecate,undeprecate}`
+- Add `buf beta registry {plugin,template} {deprecate,undeprecate}`.
+- Add warning when using enterprise dependencies without specifying a enterprise
+  remote in the module's identity.
 
 ## [v1.0.0-rc10] - 2021-12-16