+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [Unreleased]
 
 - Upgrade to `protoc` 3.19.2 support.
-- Add `buf beta registry {plugin,template} {deprecate,undeprecate}`
 - Add `PACKAGE_NO_IMPORT_CYCLE` lint rule to `OTHER` lint category to detect package import cycles.
+- Add `buf beta registry {plugin,template} {deprecate,undeprecate}`.
+- Add warning when using enterprise dependencies without specifying a enterprise
+  remote in the module's identity.
 
 ## [v1.0.0-rc10] - 2021-12-16