+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Add `buf beta registry {plugin,template} {deprecate,undeprecate}`.
 - Add warning when using enterprise dependencies without specifying a enterprise
   remote in the module's identity.
+- Remove `digest`, and `created_at` fields from the `buf.lock`. This will temporarily create a new commit
+  when pushing the same contents to an existing repository, since the `ModulePin` has been reduced down.
 - Add `--track` flag to `buf push`
 - Update `buf beta registry commit list` to allow a track to be specified.
 - Add `buf beta registry track {list,delete}` commands.